Add tests for ValueOption (! not supported)
Note: `ValueOption` is supported. Tests are just in case we want to add support later on. Tests will be removed again in next commit!

`ValueNone` isn't covered by default `null` Handling -> required custom handling.

To not emit Property with `ValueNone`:
In `IContractResolver.CreateObjectContract(objectType)` add for property of type `ValueOption` handler for `GetIsSpecified` or `ShouldSerialize`.
Issue: predicate receives not value of property, but object it's contained in.
-> Property must be extracted first before checking if `ValueNone`