Fix textDocument/prepareRename default behavior
notImplemented cannot be used because the server will fail on rename.
None cannot be used because it marks an invalid rename.

We need to respond with a {defaultBehavior=true} so that those servers
that didn't override prepareRename still have the same rename behavior
as before.

NOTE: I needed to model the response fully to make things work.