Avoid server crash when handler returns LspResult.Error
It appears that https://github.com/ionide/LanguageServerProtocol/pull/29 broke the reporting of
LspResult.Error to StreamJsonRpc.

StreamJsonRpc processes LocalRpcException to report exceptions/errors from request handlers.
In particular we have code in `requestHandling` that actually throws LocalRcpException when
handler returns LspResult.Error.

This commit adds code to skip markingLocalRcpException as a fatal exception and make it not
crash the server.